Add fileworker module and thread support
Introduced new fileworker.cpp and fileworker.h source files to the project.
I added:
- a task queue
- background worker thread
- graceful shutdown.

basically, I have implemented a queue system so that the user gets to triage photos as fast as they want, and the worker thread will move the photos in the background, so the program feels incredibly responsive.